Moved CSGBrushSubMesh into separate file
