+++ Packages/com.chisel.components/Chisel/Components/API.public/Assets/CSGBrushMeshAsset/CSGBrushSubMesh.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 90a22b9fbff655a41b62e030885db01f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
