+++ Packages/com.chisel.core/Chisel/Core/API.public/BrushMesh.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.ComponentModel;
 using System.Runtime.InteropServices;
 using UnityEngine;
